Rename relayCacheClient
