+++ src/v2/Artsy/Relay/middleware/cache/Cache.ts
@@ -1,7 +1,7 @@
 import createLogger from "v2/Utils/logger"
 import RelayQueryResponseCache from "relay-runtime/lib/network/RelayQueryResponseCache"
 import { isServer } from "lib/environment"
-import { cache } from "lib/relayCacheClient"
+import { cache } from "lib/cacheClient"
 
 import {
   GraphQLResponse,
